requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
cryptography==43.0.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
iniconfig==2.0.0
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
paramiko==3.4.1
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sshtunnel==0.4.0
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wsproto==1.2.0
